Implemented Composition API to create component functionality and import component functionality
